tools/libxl: introduce libxl__domain_restore_device_model to load qemu state
In normal migration, the qemu state was passed to qemu as a parameter.
With COLO, Secondary vm is running. So we will do the following steps
at every checkpoint:
1. suspend both primay vm and secondary vm
2. sync the state
3. resume both primary vm and secondary vm
Primary will send qemu's state in step2, and
Secondary's qemu should read it and restore the state before it
is resumed. We can not pass the state to qemu as a parameter because
Secondary QEMU already started at this point, so we introduce
libxl__domain_restore_device_model() to do it.
This API should be called before resuming secondary vm.

Signed-off-by: Yang Hongyang <yanghy@cn.fujitsu.com>
Signed-off-by: Wen Congyang <wency@cn.fujitsu.com>
Cc: Anthony Perard <anthony.perard@citrix.com>