tools/libx{l,c}: add back channel to libxc
We need to send secondary's dirty page pfns back to primary.
Primary will then send pages that are both dirtied on
primary/secondary to secondary. in this way the secondary's
memory will be consistent with primary.

We do this on libxc side, so we need to introduce back
channel to libxc.

Signed-off-by: Yang Hongyang <yanghy@cn.fujitsu.com>
CC: Ian Campbell <Ian.Campbell@citrix.com>
CC: Ian Jackson <Ian.Jackson@eu.citrix.com>
CC: Wei Liu <wei.liu2@citrix.com>
CC: Andrew Cooper <andrew.cooper3@citrix.com>